begin fixing tests from StateLanguage
