+++ tests/casper/test_view.py
@@ -2,7 +2,7 @@ import pytest
 
 from casper.abstract_view import AbstractView
 from casper.protocols.blockchain.block import Block
-from simulations.testing_language import TestLangCBC
+from simulations.blockchain_test_lang import BlockchainTestLang
 
 
 TEST_WEIGHT = {0: 10, 1: 11}
@@ -16,8 +16,8 @@ def test_new_view():
 
 
 def test_justification_stores_hash():
-    test_lang = TestLangCBC(TEST_WEIGHT)
-    test_lang.parse('B0-A S1-A B1-B')
+    test_lang = BlockchainTestLang(TEST_WEIGHT)
+    test_lang.parse('M0-A SJ1-A M1-B')
 
     validator_0 = test_lang.validator_set.get_validator_by_name(0)
     validator_1 = test_lang.validator_set.get_validator_by_name(1)
@@ -27,13 +27,13 @@ def test_justification_stores_hash():
     assert len(justification) == 2
     assert not isinstance(justification[validator_0], Block)
     assert not isinstance(justification[validator_1], Block)
-    assert justification[validator_0] == test_lang.blocks['A'].hash
-    assert justification[validator_1] == test_lang.blocks['B'].hash
+    assert justification[validator_0] == test_lang.messages['A'].hash
+    assert justification[validator_1] == test_lang.messages['B'].hash
 
 
 def test_justification_includes_justified_messages():
-    test_lang = TestLangCBC(TEST_WEIGHT)
-    test_lang.parse('B0-A B0-B P1-B B1-C')
+    test_lang = BlockchainTestLang(TEST_WEIGHT)
+    test_lang.parse('M0-A M0-B SJ1-B M1-C')
 
     validator_0 = test_lang.validator_set.get_validator_by_name(0)
     validator_1 = test_lang.validator_set.get_validator_by_name(1)
@@ -41,52 +41,52 @@ def test_justification_includes_justified_messages():
     justification = validator_1.justification()
 
     assert len(justification) == 2
-    assert test_lang.blocks["A"].hash not in justification.values()
+    assert test_lang.messages["A"].hash not in justification.values()
     assert test_lang.network.global_view.genesis_block.hash in justification.values()
-    assert justification[validator_1] == test_lang.blocks['C'].hash
+    assert justification[validator_1] == test_lang.messages['C'].hash
 
-    test_lang.parse('S1-B')
+    test_lang.parse('SJ1-B')
 
     justification = validator_1.justification()
 
     assert len(justification) == 2
-    assert justification[validator_0] == test_lang.blocks['B'].hash
-    assert justification[validator_1] == test_lang.blocks['C'].hash
+    assert justification[validator_0] == test_lang.messages['B'].hash
+    assert justification[validator_1] == test_lang.messages['C'].hash
 
 
 def test_add_justified_message():
-    test_lang = TestLangCBC(TEST_WEIGHT)
-    test_lang.parse('B0-A B0-B S1-A')
+    test_lang = BlockchainTestLang(TEST_WEIGHT)
+    test_lang.parse('M0-A M0-B SJ1-A')
     validator_0 = test_lang.validator_set.get_validator_by_name(0)
     validator_1 = test_lang.validator_set.get_validator_by_name(1)
-    assert test_lang.blocks['A'] in validator_0.view.justified_messages.values()
-    assert test_lang.blocks['A'] in validator_1.view.justified_messages.values()
-    assert test_lang.blocks['B'] in validator_0.view.justified_messages.values()
-    assert test_lang.blocks['B'] not in validator_1.view.justified_messages.values()
+    assert test_lang.messages['A'] in validator_0.view.justified_messages.values()
+    assert test_lang.messages['A'] in validator_1.view.justified_messages.values()
+    assert test_lang.messages['B'] in validator_0.view.justified_messages.values()
+    assert test_lang.messages['B'] not in validator_1.view.justified_messages.values()
 
 
 @pytest.mark.parametrize(
     'test_string, justified_messages, unjustified_messages',
     [
-        ('B0-A B0-B P1-B', [['A', 'B'], []], [[], ['B']]),
-        ('B0-A B0-B P1-B S1-A', [['A', 'B'], ['A', 'B']], [[], []]),
-        ('B0-A B0-B B0-C P1-C P1-B', [['A', 'B', 'C'], []], [[], ['B', 'C']]),
+        ('M0-A M0-B S1-B', [['A', 'B'], []], [[], ['B']]),
+        ('M0-A M0-B S1-B SJ1-A', [['A', 'B'], ['A', 'B']], [[], []]),
+        ('M0-A M0-B M0-C S1-C S1-B', [['A', 'B', 'C'], []], [[], ['B', 'C']]),
     ]
 )
 def test_only_add_justified_messages(test_string, justified_messages, unjustified_messages):
-    test_lang = TestLangCBC(TEST_WEIGHT)
+    test_lang = BlockchainTestLang(TEST_WEIGHT)
     test_lang.parse(test_string)
 
     for validator in test_lang.validator_set:
         idx = validator.name
 
         for message in justified_messages[idx]:
-            assert test_lang.blocks[message] in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] not in validator.view.pending_messages.values()
+            assert test_lang.messages[message] in validator.view.justified_messages.values()
+            assert test_lang.messages[message] not in validator.view.pending_messages.values()
 
         for message in unjustified_messages[idx]:
-            assert test_lang.blocks[message] not in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] in validator.view.pending_messages.values()
+            assert test_lang.messages[message] not in validator.view.justified_messages.values()
+            assert test_lang.messages[message] in validator.view.pending_messages.values()
 
 
 @pytest.mark.parametrize(
@@ -94,24 +94,24 @@ def test_only_add_justified_messages(test_string, justified_messages, unjustifie
     [
         (
             TEST_WEIGHT,
-            'B0-A B0-B P1-B',
+            'M0-A M0-B S1-B',
             [['A', 'B'], []],
             [[], ['B']],
-            'S1-A'
+            'SJ1-A'
         ),
         (
             {0: 10, 1: 9, 2: 8},
-            'RR0-A B0-B S1-B B1-C P2-C',
+            'RR0-A M0-B SJ1-B M1-C S2-C',
             [['A', 'B'], ['A', 'B'], []],
             [[], [], ['C']],
-            'P2-B'
+            'S2-B'
         ),
         (
             TEST_WEIGHT,
-            'B0-A S1-A B0-B B0-C B0-D B0-E P1-E',
+            'M0-A SJ1-A M0-B M0-C M0-D M0-E S1-E',
             [['A', 'B', 'C', 'D', 'E'], ['A'], []],
             [[], [], ['E']],
-            'P1-B P1-C P1-D'
+            'S1-B S1-C S1-D'
         ),
         (
             TEST_WEIGHT,
@@ -123,19 +123,19 @@ def test_only_add_justified_messages(test_string, justified_messages, unjustifie
     ]
 )
 def test_resolve_message_when_justification_arrives(weight, test_string, justified_messages, unjustified_messages, resolving_string):
-    test_lang = TestLangCBC(weight)
+    test_lang = BlockchainTestLang(weight)
     test_lang.parse(test_string)
 
     for validator in test_lang.validator_set:
         idx = validator.name
 
         for message in justified_messages[idx]:
-            assert test_lang.blocks[message] in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] not in validator.view.pending_messages.values()
+            assert test_lang.messages[message] in validator.view.justified_messages.values()
+            assert test_lang.messages[message] not in validator.view.pending_messages.values()
 
         for message in unjustified_messages[idx]:
-            assert test_lang.blocks[message] not in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] in validator.view.pending_messages.values()
+            assert test_lang.messages[message] not in validator.view.justified_messages.values()
+            assert test_lang.messages[message] in validator.view.pending_messages.values()
 
     test_lang.parse(resolving_string)
 
@@ -143,25 +143,25 @@ def test_resolve_message_when_justification_arrives(weight, test_string, justifi
         idx = validator.name
 
         for message in justified_messages[idx]:
-            assert test_lang.blocks[message] in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] not in validator.view.pending_messages.values()
+            assert test_lang.messages[message] in validator.view.justified_messages.values()
+            assert test_lang.messages[message] not in validator.view.pending_messages.values()
 
         for message in unjustified_messages[idx]:
-            assert test_lang.blocks[message] in validator.view.justified_messages.values()
-            assert test_lang.blocks[message] not in validator.view.pending_messages.values()
+            assert test_lang.messages[message] in validator.view.justified_messages.values()
+            assert test_lang.messages[message] not in validator.view.pending_messages.values()
 
 
 
 def test_multiple_messages_arriving_resolve():
     test_string = "B0-A S1-A B0-B B0-C B0-D B0-E B0-F P1-F"
-    test_lang = TestLangCBC(TEST_WEIGHT)
+    test_lang = BlockchainTestLang(TEST_WEIGHT)
     test_lang.parse(test_string)
 
     validator_1 = test_lang.validator_set.get_validator_by_name(1)
 
     assert len(validator_1.view.justified_messages) == 2
     assert len(validator_1.view.pending_messages) == 1
-    assert test_lang.blocks['F'] in validator_1.view.pending_messages.values()
+    assert test_lang.messages['F'] in validator_1.view.pending_messages.values()
 
     validator_1.receive_messages(test_lang.network.global_view.justified_messages.values())
